Made quality of life improvements
Allowed REPL to be able to SAVE and LOAD its command.
Standardized behaviour for Option and Box in interpreter.
Made Struct and Enum types work on a pointer system. Also added an InterpreterFrom trait to better accomodate this functionality.
Removed generics from ScopeErr.
Functions like get_object and get_var now expect a pointer to said var or object and not a name and scope.
Fixed ParserDataType not recognizing dyn keyword.